% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_plots.R
\name{disp_add_slide}
\alias{disp_add_slide}
\title{Add a new slide to a PowerPoint presentation and insert a chart}
\usage{
disp_add_slide(
  presentation,
  chart,
  layout = "Title and Content",
  master = "Office Theme"
)
}
\arguments{
\item{presentation}{An officer::pptx object representing the PowerPoint presentation.}

\item{chart}{An mschart::ms_chart object representing the chart to be inserted.}

\item{layout}{A character string specifying the slide layout to be used (default: "Title and Content").}

\item{master}{A character string specifying the master slide to be used (default: "Office Theme").}
}
\value{
A modified officer::pptx object with the new slide and chart added.
}
\description{
This function creates a new slide with a specified layout and inserts a specified chart into the PowerPoint presentation using the officer package.
}
