% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{disp_example}
\alias{disp_example}
\title{DisplayR example}
\format{
A data frame with 5,000 observations of 15 variables.
\describe{
\item{document}{A document id}
\item{date}{placeholder}
\item{text_copy}{placeholder}
\item{text}{placeholder}
\item{topic}{placeholder}
\item{permalink}{placeholder}
\item{cluster}{placeholder}
\item{sentiment}{placeholder}
\item{V1}{placeholder}
\item{V2}{placeholder}
\item{new_clusters}{placeholder}
\item{universal_message_id}{placeholder}
\item{fifty_clusters}{placeholder}
\item{topic_n}{placeholder}
\item{sender_screen_name}{placeholder}
}
}
\usage{
disp_example
}
\description{
Data frame with some useful variables for demonstration purposes/
}
\keyword{internal}
