% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{disp_flextable}
\alias{disp_flextable}
\title{Create a flextable for volume + sentiment with a graident fill.}
\usage{
disp_flextable(
  data,
  group_var,
  sentiment_var,
  positive_colour = "#107C10",
  negative_colour = "#D83B01"
)
}
\arguments{
\item{data}{A data frame containing the group variable and sentiment variable.}

\item{group_var}{The name of the column in the data frame used for grouping.}

\item{sentiment_var}{The name of the column in the data frame containing sentiment values.}

\item{positive_colour}{(Optional) The color for the positive gradient. Default is "#107C10".}

\item{negative_colour}{(Optional) The color for the negative gradient. Default is "#D83B01".}
}
\value{
A flextable object displaying the summary table with gradient coloring for positive and negative sentiments.
}
\description{
This function takes a data frame, the grouping variable, and the sentiment variable to create a summary table and display it as a flextable. The table contains the count and percentage of positive, negative, and neutral sentiments per group. The positive and negative columns are filled with a gradient color.
}
\details{
The output of this function can be exported directly to PowerPoint.
}
