% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_gt.R
\name{disp_gt_sent_time}
\alias{disp_gt_sent_time}
\title{Display Sentiment Over Time}
\usage{
disp_gt_sent_time(
  data,
  sentiment_var = sentiment,
  date_var = date,
  chart_type = c("lines", "bars"),
  time_unit = c("week", "day", "month", "quarter", "year"),
  date_breaks = "4 months",
  date_labels = "\%b"
)
}
\arguments{
\item{data}{The input data frame.}

\item{sentiment_var}{The name of the variable in the data frame containing the sentiment values.}

\item{date_var}{The name of the variable in the data frame containing the dates.}

\item{chart_type}{The type of chart to be plotted. Default is "lines".}

\item{time_unit}{The time unit to group the data by. Default is "week".}

\item{date_breaks}{The breaks to be used on the x-axis for the date scale. Default is "4 months".}

\item{date_labels}{The format of the labels on the x-axis for the date scale. Default is "\%b".}
}
\value{
A ggplot object representing the plotted chart.
}
\description{
This function generates a chart showing the sentiment distribution over time based on a data frame, styled for gt.
}
\keyword{internal}
