% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_gt.R
\name{disp_gt_vot}
\alias{disp_gt_vot}
\title{Create volume over time charts for use in gt tables}
\usage{
disp_gt_vot(
  df,
  date_var,
  time_unit = c("week", "day", "month", "quarter", "year"),
  bar_colour = "#628EFD",
  date_breaks = "4 months",
  date_labels = "\%b"
)
}
\arguments{
\item{df}{The input data frame.}

\item{date_var}{The name of the variable in the data frame containing the dates.}

\item{time_unit}{The time unit to group the data by. Default is "week".}

\item{bar_colour}{The color of the bars in the plot. Default is "#628EFD".}

\item{date_breaks}{The breaks to be used on the x-axis for the date scale. Default is "4 months".}

\item{date_labels}{The format of the labels on the x-axis for the date scale. Default is "\%b".}
}
\value{
A ggplot object representing the bar plot.
}
\description{
The plot this function output appears in the cells of a gt table. There is custom styling applied to remove much of the detail, the output is not
}
\keyword{internal}
