% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_plots.R
\name{disp_ms_sent_grouped}
\alias{disp_ms_sent_grouped}
\title{Display a Stacked Bar Chart of Sentiment Grouped by a Variable}
\usage{
disp_ms_sent_grouped(
  data,
  sentiment_var,
  group_var,
  plot_type = c("percent", "volume")
)
}
\arguments{
\item{data}{A data frame containing the sentiment and group variables.}

\item{sentiment_var}{The name of the sentiment variable in the data frame (as a string or unquoted symbol).}

\item{group_var}{The name of the group variable in the data frame (as a string or unquoted symbol).}

\item{plot_type}{Where raw volumes or percentages should be plotted.}
}
\value{
A \code{mschart::ms_barchart} object representing the stacked bar chart.
}
\description{
This function takes a data frame, a sentiment variable, and a group variable, and creates a stacked bar chart representing the distribution of sentiment. within each group, with the bars displayed horizontally and the percentage of each sentiment shown.
}
