% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayR.R
\name{displayRobfuscated}
\alias{displayRobfuscated}
\title{Display R}
\usage{
displayRobfuscated()
}
\description{
Displays a capital R. Same outcome as displayR() but the code
has be obfuscated. That isn't to say that the original code
is very clear...
}
\examples{
displayRobfuscated()
}

