% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brands.R
\name{dr_list_brands}
\alias{dr_list_brands}
\title{View which brands have styling options available for them}
\usage{
dr_list_brands(quiet = FALSE)
}
\arguments{
\item{quiet}{whether to print the name of detected brands or not.}
}
\value{
A list of folder names relating to brands/
}
\description{
Check which brands a _brand.yml and an optional 'logo.png' have been uploaded to DisplayR for.
}
\examples{
brands <- dr_list_brands()
}
