% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppt_plots.R
\name{dr_mmlb_barchart}
\alias{dr_mmlb_barchart}
\title{Create a stacked bar chart for Microsoft's Making My Life Better classification}
\usage{
dr_mmlb_barchart(data, colours, state_var)
}
\arguments{
\item{data}{A dataframe containing the states and their corresponding 'yes' and 'no' counts.}

\item{colours}{A named vector of colors to apply to the 'yes' groups in the chart.}

\item{state_var}{The column with the state being plotted (x axis)
#'}
}
\value{
An object of class 'ms_barchart' representing the stacked bar chart.
}
\description{
Generates a stacked bar chart using 'mschart' with custom colors for 'yes' and 'no' groups.
}
\examples{
df <- dplyr::tribble(
  ~state, ~yes, ~no,
  "Brings me joy", 4861, 167,
  "Empowers me", 4715, 313,
  "Feels connected", 922, 4106,
  "Inspires new ideas", 4906, 122,
  "Simplifies tech", 3954, 1074
)

colours <- c("#0078D4", "#107C10", "#FFB900", "#008575", "#D83B01")
chart <- dr_mmlb_barchart(df, colours, state_var = state)
print(chart, preview = TRUE)

}
