% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_sentiment.R
\name{dr_plot_sent}
\alias{dr_plot_sent}
\title{Create a sentiment distribution bar chart}
\usage{
dr_plot_sent(
  data,
  sentiment_var = sentiment,
  bar_labels = c("percent", "volume", "none"),
  sentiment_colours = c(positive = "#107C10", negative = "#D83B01", neutral = "#FFB900")
)
}
\arguments{
\item{data}{Data Frame or Tibble object}

\item{sentiment_var}{Variable containing 'positive', 'negative', 'neutral' sentiment classification}

\item{bar_labels}{Data labels for the bars, percent or volume}

\item{sentiment_colours}{Colour mapping for the sentiment categories}
}
\value{
ggplot object of sentiment distribution
}
\description{
We prefer bars over pie charts or donuts, end.
}
