% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_volume.R
\name{dr_plot_vot}
\alias{dr_plot_vot}
\title{Quickly create a bar chart which shows volume over time.}
\usage{
dr_plot_vot(
  data,
  date_var = date,
  plot_type = c("line", "bar"),
  colour = "#440154FF",
  time_unit = c("day", "week", "month", "quarter", "year"),
  smooth = FALSE
)
}
\arguments{
\item{data}{Data Frame or Tibble object}

\item{date_var}{Date column}

\item{plot_type}{Should the plot be shown as a line or a bar chart?}

\item{colour}{Colour of line/bars - string, name of hexcode}

\item{time_unit}{select unit e.g. ("day", "week", "month", "quarter", "year")}

\item{smooth}{Add a smoothing line (geom_smooth) to the plot?}
}
\value{
ggplot object
}
\description{
You can use the \code{time_unit} argument to display counts for different time frames, e.g. \code{time = "day"} will give a chart where each bar represents the volume counts for a day. To give the bars new colours, enter a string into \verb{bar_colour =  }, the string should be a hexcode e.g. "#440154FF" or the name of a colour e.g. "midnightblue"
}
\examples{
DisplayR::disp_example \%>\% disp_plot_vot(date = date, time_unit = "day")
DisplayR::disp_example \%>\% disp_plot_vot(date = date, colour = "midnightblue", time_unit = "week")
}
