% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_volume.R
\name{dr_plot_vot_group}
\alias{dr_plot_vot_group}
\title{Quickly plot faceted volume of groups over time}
\usage{
dr_plot_vot_group(
  data,
  group_var,
  date_var = date,
  time_unit = c("day", "week", "month", "quarter", "year"),
  nrow = 2
)
}
\arguments{
\item{data}{Data frame or tibble}

\item{group_var}{grouping variable e.g. country, cluster, topic etc.}

\item{date_var}{Variable which contains date information (can be datetime too I think)}

\item{time_unit}{A single unit of time fed into lubridate::floor_date  "week", "day", "month","quarter", "year"}

\item{nrow}{How many rows the plot should be shown in}
}
\value{
ggplot object of facetted bar charts
}
\description{
Quickly plot faceted volume of groups over time
}
