% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_gt.R
\name{make_gt_summary_table}
\alias{make_gt_summary_table}
\title{Create a Summary Table for gt Summary Process}
\usage{
make_gt_summary_table(data, group_var, sentiment_var)
}
\arguments{
\item{data}{The input data frame.}

\item{group_var}{The name of the variable used for grouping the data.}

\item{sentiment_var}{The name of the variable representing the sentiment values.}
}
\value{
A data frame containing the summary table.
}
\description{
This function creates a summary table as one step of the gt summary process. The table summarizes the data based on a group variable and a sentiment variable.
}
\keyword{internal}
