% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_samy_continuous}
\alias{theme_samy_continuous}
\title{theme_samy_continuous}
\usage{
theme_samy_continuous(
  index = NULL,
  direction = 1,
  guide = "legend",
  font_family = "Montserrat Regular"
)
}
\arguments{
\item{index}{Choose palettes colours by index by setting index equal to a character vector e.g. c(1,2,3) or c(1:3)}

\item{direction}{The direction of the colours in the scale. Set to -1 to reverse them.}

\item{guide}{The type of legend. Use "colourbar", "legend" or FALSE.}
}
\description{
Adds Samy colours and font to continuous plot.
}
\keyword{internal}
