% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_volume.R
\name{dr_plot_vot_group}
\alias{dr_plot_vot_group}
\title{Quickly make a line plot of volume over time for a grouping variable}
\usage{
dr_plot_vot_group(
  data,
  group_var,
  date_var = date,
  time_unit = c("day", "week", "month", "quarter", "year")
)
}
\arguments{
\item{data}{Data frame or tibble}

\item{group_var}{grouping variable e.g. country, cluster, topic etc.}

\item{date_var}{Variable which contains date information (can be datetime too I think)}

\item{time_unit}{A single unit of time fed into lubridate::floor_date  "week", "day", "month","quarter", "year"}
}
\value{
ggplot object
}
\description{
Quickly make a line plot of volume over time for a grouping variable
}
