% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_capture_continuous}
\alias{theme_capture_continuous}
\title{theme_capture_continuous}
\usage{
theme_capture_continuous(
  direction = c("forwards", "backwards"),
  guide = c("colourbar", "colorbar", "legend", "none"),
  font_family = "GT Walsheim Pro"
)
}
\arguments{
\item{direction}{The direction of the colours in the scale. Set to -1 to reverse them.}

\item{guide}{The type of legend. Use "colourbar", "legend" or FALSE.}
}
\description{
Adds Capture colours and font to continuous plot.
}
\keyword{internal}
