% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_capture_discrete}
\alias{theme_capture_discrete}
\title{theme_capture_discrete}
\usage{
theme_capture_discrete(direction = 1, font_family = "GT Walsheim Pro")
}
\arguments{
\item{direction}{The direction of the colours in the scale. Set to -1 to reverse them.}
}
\description{
Adds Capture colours and font to discrete plot.
}
\keyword{internal}
