% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_samy_discrete}
\alias{theme_samy_discrete}
\title{theme_samy_discrete}
\usage{
theme_samy_discrete(index = NULL, font_family = "Montserrat Regular")
}
\arguments{
\item{index}{Choose palettes colours by index by setting index equal to a character vector e.g. c(1,2,3) or c(1:3)}
}
\description{
Adds Samy colours and font to discrete plot.
}
\keyword{internal}
