% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{vot_unit_data}
\alias{vot_unit_data}
\title{Get unit-specific data for plotting and switch o reduce complexity}
\usage{
vot_unit_data(time_unit, vot_variable = "Volume of Mentions", unit = "Count")
}
\arguments{
\item{time_unit}{A character string specifying the time unit.
One of "day", "week", "month", "quarter", or "year".}

\item{unit}{A character string used to describe the unit of the data in the y-axis label.
Default is "Count".}

\item{variable}{A character string used to describe the data in the plot title.
Default is "Volume of Mentions".}
}
\value{
A list with elements 'date_breaks', 'date_labels', 'title', 'yaxis'
}
\description{
This function returns a list of values used for configuring the ggplot2 plot
depending on the time unit. This includes date breaks, date labels, plot title,
and y-axis label. The 'variable' and 'unit' parameters allow for flexibility in
describing the data being plotted.
}
\examples{
get_unit_data(time_unit = "day")
get_unit_data(time_unit = "week", variable = "Sentiment Analysis", unit = "Score")
}
